pyproject.toml
[tool.poetry]
name = "cocktail-cli"
version = "0.1.2"
authors = ["0.1.0 <mat.p.wasilewski@gmail.com>"]
description = "CLI for cocktail makers!"

[tool.poetry.scripts]
cocktail = "cocktail_cli.cli:cli"

[tool.poetry.dependencies]
python = "^3.7"
JSON-log-formatter = "0.4.0"
ujson = "4.3.0"
pyjwt = "2.3.0"
types-python-slugify = "5.0.2"
types-ujson = "4.2.0"
types-setuptools = "57.4.4"
types-requests = "2.26.1"
tomlkit = "^0.7.2"
pydantic = "^1.8.2"
click = "^8.1.3"
requests-mock = "^1.9.3"
requests-cache = "^0.9.5"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
pytest-cov = "3.0.0"
pytest-asyncio = "0.16.0"
black = "21.12b0"
mypy = "0.910"
isort = "5.10.1"
autoflake = "1.4"
sqlalchemy-stubs = "0.4"
flakehell = "0.9.0"
flake8 = "3.9"
flake8-commas = "2.1.0"
flake8-quotes = "3.3.1"
testfixtures = "6.18.3"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 79
use_parentheses = true
ensure_newline_before_comments = true

[tool.flakehell]
extended_default_ignore = []
exclude = ["README.md"]
format = "colored"
max_line_length = 79
show_source = true

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-darglint = ["+*"]
flake8-docstrings = ["+*"]
flake8-eradicate = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
flake8-spellcheck = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*"]
pyflakes = ["+*"]
pylint = ["+*"]

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
